+++ app/src/main/java/com/example/allgasnobrakes/MainActivity.java
@@ -39,11 +39,11 @@ public class MainActivity extends AppCompatActivity {
         to register
          */
         if (savedInstanceState == null) {
-//            String id = Settings.Secure.getString(getContentResolver(), Settings.Secure.ANDROID_ID);
-            String id = "DAJ101";
+            String id = Settings.Secure.getString(getContentResolver(), Settings.Secure.ANDROID_ID);
+//            String id = "DAJ101";
 
             FirebaseFirestore db = FirebaseFirestore.getInstance();
-            db.useEmulator("10.0.2.2", 8080);
+//            db.useEmulator("10.0.2.2", 8080);
             DocumentReference cloudID = db.collection("DeviceID").document(id);
 
             //https://firebase.google.com/docs/firestore/query-data/get-data
