+++ app/src/main/res/layout/scanner.xml
@@ -88,7 +88,7 @@
         android:alpha="0"
         android:textAlignment="center"
         android:textColor="@color/white"
-        android:textSize="30sp"
+        android:textSize="24sp"
         android:text="TextView" />
 
 </RelativeLayout>
